backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
torch==2.0.1
torchvision==0.15.1
transformers==4.30.0
pillow==9.5.0
